result_ms/requirements.txt
pip==9.0.2
appdirs==1.4.3
Django==1.11
djangorestframework==3.6.2
mysqlclient
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
